Fixes bug caused by babel helpers being overriden by each bundle file
If more than one bundle from babel-globals is added on a page, the babel helpers of the last added bundle would be the only available, since they were overriding each other's global variable.
This changes babel helpers to be only a local var instead of a global, which prevents this kind of conflict.