Adds option for opting out of setting exported modules inside `this` - Fixes #4
babel-globals forces the exported modules to be included in this by default. If storeOnThis is set to false now though, this won't happen. This means that users will have full control over where the module variables will be exported to, since they can just set `globalName` to whatever they want without us forcing that to be included in `this`.